backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.28.0
numpy>=1.26.0
rank-bm25>=0.2.2
sentence-transformers>=2.5.1
